refactor: remove unused dependency; update imports to reflect new data structure
