+++ app/[domain]/page.tsx
@@ -1,7 +1,7 @@
 import { UsaUniversityCountdown } from "@/components/UsaUniversityCountdown";
 import { notFound } from "next/navigation";
 import type { Metadata } from "next";
-import universities from "@/universities";
+import universities from "@/data/universities";
 import type { Viewport } from "next";
 
 interface Props {
