--- src/types/global.d.ts
@@ -1,4 +1,0 @@
-interface Window {
-  dataLayer: unknown[];
-  gtag: (command: string, ...args: unknown[]) => void;
-}
